requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
psycopg[binary]==3.2.6
pydantic-settings==2.8.1
python-multipart==0.0.20
httpx==0.28.1
playwright==1.51.0
rapidfuzz==3.13.0
